third_party/WebKit/LayoutTests/fast/css/fontface-arraybuffer-expected.txt

CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
Tests ArrayBuffer / ArrayBufferView constructors of FontFace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS face1.status is "loaded"
PASS face2.status is "loaded"
PASS face3.status is "error"
PASS rejectionValue.name is "SyntaxError"
PASS document.getElementById('FontFromArrayBuffer').offsetWidth is document.getElementById('ref').offsetWidth
PASS document.getElementById('FontFromArrayBufferView').offsetWidth is document.getElementById('ref').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
abc
abc
abc
